Renamed format method; Moved the formatter to utility class
